fix(voting): atomic rate-limit increment and shared assignRanks helper
- Switch submission rate-limit from read-then-increment to atomic incr+check
  so parallel requests from the same cookie cannot bypass the limit.
- Extract assignRanks to src/lib/voting.ts and import it from the admin
  results page and API to eliminate duplicated tie-ranking logic.
