+++ src/app/admin/results/page.tsx
@@ -2,6 +2,7 @@ import { cookies } from 'next/headers';
 import { redirect } from 'next/navigation';
 import { ADMIN_SESSION_COOKIE, verifySession } from '@/lib/admin';
 import {
+  assignRanks,
   getEventConfig,
   isWithinWindow,
   listBallots,
@@ -9,25 +10,9 @@ import {
   tallyVotes,
   voterCount,
 } from '@/lib/voting';
-import type { Submission } from '@/types/voting';
 import { ResultsExport } from './ResultsExport';
 
 export const dynamic = 'force-dynamic';
-
-function assignRanks(rows: Array<Submission & { voteCount: number }>) {
-  const sorted = [...rows].sort((a, b) => b.voteCount - a.voteCount);
-  let rank = 0;
-  let lastCount = -1;
-  let seen = 0;
-  return sorted.map((r) => {
-    seen += 1;
-    if (r.voteCount !== lastCount) {
-      rank = seen;
-      lastCount = r.voteCount;
-    }
-    return { ...r, rank };
-  });
-}
 
 export default async function AdminResultsPage() {
   const session = cookies().get(ADMIN_SESSION_COOKIE)?.value;
